unifying data type to 32

+++ docs/tutorials/hssm-tst-gpu.py
@@ -223,11 +223,15 @@ def main():
     dataset.rename(columns={'rt1': 'rt'}, inplace=True)
 
     dataset["rt"] = dataset["rt"].astype('float32')
-    dataset["response"] = dataset["response"].astype('int32')
+    dataset["response"] = dataset["response"].astype('int32')   
+    dataset["feedback"] = dataset["feedback"].astype('int32')
+    dataset["valid_upto"] = dataset["valid_upto"].astype('int32')
+    dataset["trial_id"] = dataset["trial_id"].astype('int32')
+    dataset["state1"] = dataset["state1"].astype('int32')
+    dataset["state2"] = dataset["state2"].astype('int32')
+    dataset["response2"] = dataset["response2"].astype('int32')
+    dataset["participant_id"] = dataset["participant_id"].astype('int32')
 
-    dataset['state1']=dataset['state1'].astype('int32')
-    dataset['state2']=dataset['state2'].astype('int32')
-    dataset['participant_id']=dataset['participant_id'].astype('int32')
 
     dataset = add_valid_upto_and_pad(dataset)
